refactor(wave4-step2): move lockfile/cache logic behind stable facades
