+++ crates/assay-registry/src/cache_next/keys.rs
@@ -2,3 +2,9 @@
 //!
 //! Planned ownership (Step2+):
 //! - cache key and path derivation helpers
+
+use std::path::{Path, PathBuf};
+
+pub(crate) fn pack_dir_impl(cache_dir: &Path, name: &str, version: &str) -> PathBuf {
+    cache_dir.join(name).join(version)
+}
